# domain-core-configuration-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/configuration/reference-data/{entityType}/{entityId}/deactivate:
    post:
      tags:
      - Reference Data Orchestration
      summary: Deactivate reference-data entity
      description: Runs ActivateReferenceEntitySaga with targetStatus=INACTIVE; fails
        with 409 when cross-entity references block deactivation.
      operationId: deactivate
      parameters:
      - name: entityType
        in: path
        required: true
        schema:
          type: string
      - name: entityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Deactivated
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "400":
          description: Invalid entity type
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "404":
          description: Entity not found
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "409":
          description: Active references block deactivation
          content:
            '*/*':
              schema:
                type: string
                format: uuid
  /api/v1/configuration/reference-data/{entityType}/{entityId}/activate:
    post:
      tags:
      - Reference Data Orchestration
      summary: Activate reference-data entity
      description: Runs ActivateReferenceEntitySaga with targetStatus=ACTIVE.
      operationId: activate
      parameters:
      - name: entityType
        in: path
        required: true
        schema:
          type: string
      - name: entityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Activated
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "400":
          description: Invalid entity type
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "404":
          description: Entity not found
          content:
            '*/*':
              schema:
                type: string
                format: uuid
  /api/v1/configuration/reference-data/updates:
    post:
      tags:
      - Reference Data Orchestration
      summary: Update reference-data entity with compensation
      description: Runs UpdateReferenceDataSaga to atomically update the entity and
        optional localizations.
      operationId: update_26
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReferenceDataSagaRequest"
        required: true
      responses:
        "200":
          description: Update completed
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "400":
          description: Invalid entity type or payload
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "404":
          description: Entity not found
          content:
            '*/*':
              schema:
                type: string
                format: uuid
        "409":
          description: Conflicting state
          content:
            '*/*':
              schema:
                type: string
                format: uuid
...
  /api/v1/configuration/tenants/{tenantId}/snapshot:
    get:
      tags:
      - Tenant Snapshot
      summary: Get tenant master-data snapshot
      description: "Aggregates countries, currencies, legal forms, language locales\
        \ and active lookup items for the tenant."
      operationId: getSnapshot
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Snapshot assembled
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantSnapshotResponse"
        "400":
          description: Invalid tenant identifier
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantSnapshotResponse"
        "500":
          description: Unexpected error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantSnapshotResponse"
...
components:
  schemas:
...
    LookupItemDto:
      required:
      - domainId
      - itemCode
      - itemLabelDefault
      type: object
      properties:
        itemId:
          type: string
          format: uuid
        domainId:
          type: string
          format: uuid
        itemCode:
          minLength: 1
          type: string
        itemLabelDefault:
          minLength: 1
          type: string
        itemDesc:
          type: string
        parentItemId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          format: int32
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
        isCurrent:
          type: boolean
        extraJson:
          type: string
        tenantId:
          type: string
          format: uuid
...
    LegalFormConfigDto:
      required:
      - code
      - name
      type: object
      properties:
        legalFormId:
          type: string
          format: uuid
        countryId:
          type: string
          format: uuid
        code:
          minLength: 1
          type: string
        name:
          minLength: 1
          type: string
        description:
          type: string
        isCorporate:
          type: boolean
        entityType:
          type: string
    LanguageLocaleDto:
      required:
      - languageCode
      - localeCode
      type: object
      properties:
        localeId:
          type: string
          format: uuid
        languageCode:
          minLength: 1
          type: string
        countryCode:
          type: string
        localeCode:
          minLength: 1
          type: string
        languageName:
          type: string
        nativeName:
          type: string
        regionName:
          type: string
        rtl:
          type: boolean
        sortOrder:
          type: integer
          format: int32
...
    CurrencyConfigDto:
      required:
      - currencyName
      - isoCode
      type: object
      properties:
        currencyId:
          type: string
          format: uuid
        isoCode:
          maxLength: 3
          minLength: 3
          type: string
        currencyName:
          minLength: 1
          type: string
        symbol:
          type: string
        decimalPrecision:
          type: integer
          format: int32
        isMajor:
          type: boolean
    CountryConfigDto:
      required:
      - countryName
      - isoCode
      type: object
      properties:
        countryId:
          type: string
          format: uuid
        isoCode:
          maxLength: 3
          minLength: 2
          type: string
        countryName:
          minLength: 1
          type: string
        svgFlag:
          type: string
...
    LocalizationPayload:
      required:
      - languageLocaleId
      - localeTag
      type: object
      properties:
        localizationId:
          type: string
          format: uuid
        languageLocaleId:
          type: string
          format: uuid
        localeTag:
          minLength: 1
          type: string
        translatedFields:
          type: object
          additionalProperties:
            type: string
    UpdateReferenceDataSagaRequest:
      required:
      - entityId
      - entityType
      type: object
      properties:
        entityType:
          minLength: 1
          type: string
        entityId:
          type: string
          format: uuid
        payload:
          type: object
        localizations:
          type: array
          items:
            $ref: "#/components/schemas/LocalizationPayload"
    TenantSnapshotResponse:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        countries:
          type: array
          items:
            $ref: "#/components/schemas/CountryConfigDto"
        currencies:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyConfigDto"
        legalForms:
          type: array
          items:
            $ref: "#/components/schemas/LegalFormConfigDto"
        languageLocales:
          type: array
          items:
            $ref: "#/components/schemas/LanguageLocaleDto"
        lookupItems:
          type: array
          items:
            $ref: "#/components/schemas/LookupItemDto"
...